requirements.txt
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
decorator==5.1.1
executing==0.10.0
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
six==1.16.0
stack-data==0.4.0
traitlets==5.3.0
wcwidth==0.2.5
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
decorator==5.1.1
executing==0.10.0
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
numpy==1.23.2
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
stack-data==0.4.0
traitlets==5.3.0
wcwidth==0.2.5
